pyproject.toml
[tool.poetry]
name = "pastebin"
version = "0.1.0"
description = ""
authors = ["Dmitrii <dmitriy.site1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
amqp = "5.2.0"
asgiref = "3.7.2"
billiard = "4.2.0"
cachetools = "5.3.2"
celery = "5.3.6"
certifi = "2024.2.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.3.0"
cryptography = "42.0.5"
defusedxml = "0.8.0rc2"
django = "5.0.1"
django-admin-interface = "0.28.6"
django-cachalot = "2.6.2"
django-colorfield = "0.11.0"
django-debug-toolbar = "4.3.0"
django-recaptcha = "4.0.0"
django-redis = "5.4.0"
flower = "2.0.1"
google-api-core = "2.17.1"
google-api-python-client = "2.118.0"
google-auth = "2.28.0"
google-auth-httplib2 = "0.2.0"
googleapis-common-protos = "1.62.0"
gunicorn = "21.2.0"
httplib2 = "0.22.0"
humanize = "4.9.0"
idna = "3.6"
kombu = "5.3.5"
oauthlib = "3.2.2"
packaging = "23.2"
pillow = "10.2.0"
prometheus-client = "0.20.0"
prompt-toolkit = "3.0.43"
protobuf = "4.25.3"
psycopg2-binary = "2.9.9"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pycparser = "2.21"
pyjwt = "2.8.0"
pyparsing = "3.1.1"
python-dateutil = "2.8.2"
python-slugify = "8.0.4"
python3-openid = "3.2.0"
pytz = "2024.1"
redis = "5.0.1"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
rsa = "4.9"
six = "1.16.0"
social-auth-app-django = "5.4.0"
social-auth-core = "4.5.3"
sqlparse = "0.4.4"
text-unidecode = "1.3"
tornado = "6.4"
tzdata = "2024.1"
uritemplate = "4.1.1"
urllib3 = "2.2.1"
vine = "5.1.0"
wcwidth = "0.2.13"
isort = "^5.13.2"
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ["py39"]

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 79